Move common operations to separate files
